Crate linux_embedded_hal

Implementation of embedded-hal traits for Linux devices

Drivers

This crate lets you use a bunch of platform agnostic drivers that are based on the embedded-hal traits. You can find them on crates.io by searching for the embedded-hal keyword.

Re-exports

pub extern crate gpio_cdev;

pub extern crate i2cdev;

pub extern crate nb;

pub extern crate serial_core;

pub extern crate serial_unix;

pub extern crate spidev;

pub extern crate sysfs_gpio;

Structs

CdevPin

Cdev pin re-export Newtype around gpio_cdev::LineHandle that implements the embedded-hal traits

Delay

Empty struct that provides delay functionality on top of thread::sleep

I2cdev

Newtype around i2cdev::linux::LinuxI2CDevice that implements the embedded-hal traits

Pin

Newtype around sysfs_gpio::Pin that implements the embedded-hal traits

Serial

Newtype around serial_unix::TTYPort that implements the embedded-hal traits.

Spidev

Newtype around spidev::Spidev that implements the embedded-hal traits

SysTimer

A periodic timer based on std::time::Instant, which is a monotonically nondecreasing clock.

SysfsPin

Sysfs pin re-export Newtype around sysfs_gpio::Pin that implements the embedded-hal traits

Enums

SpiOperation

Operation for transactional SPI trait