Crate linux_embedded_hal

Implementation of embedded-hal traits for Linux devices

Drivers

This crate lets you use a bunch of platform agnostic drivers that are based on the embedded-hal traits. You can find them on crates.io by searching for the embedded-hal keyword.

Re-exports

pub use i2cdev;

pub use nb;

pub use serialport;

pub use spidev;

pub use sysfs_gpio;

pub use gpio_cdev;

Structs

CdevPin

Cdev pin re-export Newtype around gpio_cdev::LineHandle that implements the embedded-hal traits

CdevPinError

Cdev pin re-export Error type wrapping gpio_cdev::errors::Error to implement embedded_hal::digital::Error

Delay

Empty struct that provides delay functionality on top of thread::sleep

I2CError

Error type wrapping LinuxI2CError to implement embedded_hal::i2c::ErrorKind

I2cdev

Newtype around i2cdev::linux::LinuxI2CDevice that implements the embedded-hal traits

SPIError

Error type wrapping io::Error to implement embedded_hal::spi::ErrorKind

Serial

Newtype around serialport::TTYPort that implements the embedded-hal traits.

SerialError

Error type wrapping io::ErrorKind to implement [embedded_hal::serial::ErrorKind]

SpidevBus

Spidev wrapper providing the embedded-hal SpiBus trait.

SpidevDevice

Spidev wrapper providing the embedded-hal SpiDevice trait.

SysTimer

A periodic timer based on std::time::Instant, which is a monotonically nondecreasing clock.

SysfsPin

Sysfs pin re-export Newtype around sysfs_gpio::Pin that implements the embedded-hal traits

SysfsPinError

Sysfs pin re-export Error type wrapping sysfs_gpio::Error to implement embedded_hal::digital::Error

Traits

CountDown

A count down timer

Periodic

Marker trait that indicates that a timer is periodic